+++ tests/integration/centralized.rs
@@ -53,6 +53,7 @@ mod other {
   centralized_test!(other, should_support_options_with_pipeline);
   centralized_test!(other, should_reuse_pipeline);
   centralized_test!(other, should_manually_connect_twice);
+  centralized_test!(other, should_support_options_with_trx);
 
   //#[cfg(feature = "dns")]
   // centralized_test!(other, should_use_trust_dns);