+++ tests/integration/clustered.rs
@@ -57,6 +57,7 @@ mod other {
   cluster_test!(other, should_support_options_with_pipeline);
   cluster_test!(other, should_reuse_pipeline);
   cluster_test!(other, should_manually_connect_twice);
+  cluster_test!(other, should_support_options_with_trx);
 
   //#[cfg(feature = "dns")]
   // cluster_test!(other, should_use_trust_dns);